Mise en place su patch sur Client et sur utilisateur, Fait avec les DTO

+++ src/main/java/com/projet/equipement/repository/ProduitRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProduitRepository extends JpaRepository<Produit, Integer> {
+public interface ProduitRepository extends JpaRepository<Produit,  Long> {
 
 }
